Attempting to remove packs/public (check)
Attempting to remove packs/public but running into issues with webpacker when removing main() in public.js

Deleting this unused main() [the import of MediaContainer] or MediaContainer or the Poll import in MediaContainer causes this error.
Commenting out or putting return; works fine. but cannot delete that import or reactComponents variable for some reason.